refactor pipeline_modules tests
