+++ tests/workflows/pipeline_modules/test_segmentation.py
@@ -26,29 +26,13 @@ from ophys_etl.workflows.pipeline_modules.segmentation import \
     SegmentationModule  # noqa E402
 from ophys_etl.workflows.well_known_file_types import WellKnownFileType # noqa E402
 from ophys_etl.workflows.workflow_steps import WorkflowStep # noqa E402
+from ophys_etl.test_utils.db_utils import MockSQLiteDB
 
-
-class TestSegmentation:
-    @classmethod
-    def setup_class(cls):
-        cls._tmp_dir = Path(tempfile.TemporaryDirectory().name)
-        cls._db_path = cls._tmp_dir / 'app.db'
-        os.makedirs(cls._db_path.parent, exist_ok=True)
-
-        db_url = f'sqlite:///{cls._db_path}'
-        cls._engine = IntializeDBRunner(
-            input_data={
-                'db_url': db_url
-            },
-            args=[]).run()
-        cls._rois_path = \
-            Path(__file__).parent / 'resources' / 'rois.json'
-
-    @classmethod
-    def teardown_class(cls):
-        shutil.rmtree(cls._tmp_dir)
+class TestSegmentation(MockSQLiteDB):
 
     def test_save_metadata_to_db(self):
+        _rois_path = \
+            Path(__file__).parent / 'resources' / 'rois.json'
         with Session(self._engine) as session:
             save_job_run_to_db(
                 workflow_step_name=WorkflowStep.SEGMENTATION,
@@ -57,7 +41,7 @@ class TestSegmentation:
                 module_outputs=[OutputFile(
                         well_known_file_type=(
                             WellKnownFileType.OPHYS_ROIS),
-                        path=self._rois_path
+                        path=_rois_path
                     )
                 ],
                 ophys_experiment_id='1',
